fix category small tips

+++ site/app/Http/Controllers/CategoryController.php
@@ -5,7 +5,6 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Model\Category;
 use App\Http\Requests;
-use Illuminate\Support\Facades\Storage;
 
 
 class CategoryController extends Controller
@@ -50,20 +49,16 @@ class CategoryController extends Controller
         $category->name = $request->name;
         $category->alias_name = $request->alias_name;
         $category->description = $request->description;
-        $category->cover = $request->cover?:'';
         $category->order = $request->order;
         $category->options = json_encode($request->options);
 
-        /*
         if (!$request->hasFile('cover')) {
             $category->cover = 'default.jpg';
         } elseif ($request->hasFile('cover') && $request->file('cover')->isValid()) {
-            $category->cover = $request->cover->store(date("Y") . '/' . date("m"));
+            $category->cover = $request->cover->store('cover/'.date("Y") . '/' . date("m"));
         } else {
             return redirect()->back()->withInput()->withErrors('图片有误！');
         }
-        */
-
         if ($category->save()) {
             return redirect()->back()->withInput()->withErrors('保存成功！');
         } else {
@@ -98,15 +93,17 @@ class CategoryController extends Controller
         $category->name = $request->name;
         $category->alias_name = $request->alias_name;
         $category->description = $request->description;
-        $category->cover = $request->cover;
+
         $category->order = $request->order;
         $category->options = json_encode($request->options);
-        /*
+
         if ($request->hasFile('cover') && $request->file('cover')->isValid()) {
-            Storage::delete($category->cover);
+            if($category->cover != 'default.jpg'){
+                Storage::delete($category->cover);
+            }
             $category->cover = $request->cover->store(date("Y") . '/' . date("m"));
         }
-        */
+
         if ($category->save()) {
             return redirect()->back()->withInput()->withErrors('保存成功！');
         } else {
